+++ _infra/helm/locust/tasks/collection-exercise-config.json
@@ -1,5 +1,5 @@
 {
-    "inputFiles": ["CE_Details_for_loading_into_RM.csv"],
+    "inputFiles": ["locust-tasks/CE_Details_for_loading_into_RM.csv"],
     "columnMappings": {
         "Survey_Id": "surveyRef",
         "Survey_Name": "name",
