+++ _infra/helm/locust/values.yaml
@@ -1,9 +1,11 @@
 locust:
   master:
     config:
-      target-host: "http://frontstage.minikube.svc.cluster.local:9000"
-      collection-instrument: "http://collection-instrument.minikube.svc.cluster.local:8080"
-      collection-exercise: "http://collection-exercise.minikube.svc.cluster.local:8080"
+      target-host: "http://frontstage.crockerm.svc.cluster.local:9000"
+      collection-instrument: "http://collection-instrument.crockerm.svc.cluster.local:8080"
+      collection-exercise: "http://collection-exercise.crockerm.svc.cluster.local:8080"
+      security-user-name: "admin"
+      security-user-password: "secret"
 
   worker:
     config:
